test(react): fold redirect-on-mount stall repro into state suite
Move the standalone redirect-stall.test.tsx into state.test.tsx under
'reactive context' and trim it to the essential reproduction. Same guard
(fails without the watch fix), one fewer test file.
